+++ lib/superform/rails/components/base.rb
@@ -6,9 +6,16 @@ module Superform
 
         delegate :dom, to: :field
 
-        def initialize(field, attributes: {})
+        def initialize(field, attributes: nil, **attributes_kwargs)
           @field = field
-          @attributes = attributes
+          if attributes
+            warn "[DEPRECATION] Passing `attributes:` keyword to #{self.class.name} is deprecated. " \
+                 "Pass HTML attributes as keyword arguments directly instead: " \
+                 "#{self.class.name}.new(field, **attributes)"
+            @attributes = attributes.merge(attributes_kwargs)
+          else
+            @attributes = attributes_kwargs
+          end
         end
 
         def field_attributes